Minor UI fix
